Moved listeners to inner classes
